pyproject.toml
[project]
name = "multi-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.8.0",
    "beautifulsoup4>=4.12.0",
    "duckduckgo-search>=4.0.0",
    "langchain-community>=0.0.20",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.0.40",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "structlog>=23.0.0",
]
